Lambda-functions/lambda-orchestrator/requirements.txt
langchain-aws==0.2.10
langgraph==0.2.62
python-dotenv==1.0.1